Add parentheses to third-from-last commit() call
If the code at 364 was intended to commit the menuItem1 added at 361, then it requires ending parentheses.  Without these parentheses, a commit does not occur here.

Otherwise, it's unclear whether L360 menuItem1 is added to the database, or whether it is replaced by L366 menuItem1 in the staging area before being committed.