Set Urban's veggie burger ID to 8
As written, this Python script gives Urban two veggie burgers.  As the ID's are assigned auto-increment, the IDs of Urban's two(?) veggie burgers are 1 and 9 respectively.

This does not match Udacity's lecture video, which proceeds under the assumption that Urban's (single) veggie burger has the id of 8.  This causes unnecessary confusion and detracts from the lecture.