Clean Auntie Ann's restaurant name string
As written, the name string returned for Auntie Ann's Diner is u"Auntie Ann's Diner' ".

This entry specifically is mentioned in the Udacity lecture videos.  The unexpected extra single quote and space in this string make it less intuitive to write comparison checks to ensure that the desired result was returned.

Hence, delete the trailing single quote and space.