Add New Item link generation
The video at L3-11 (https://classroom.udacity.com/courses/ud088/lessons/3593308717/concepts/36245586240923) shows a link on the main restaurant page to create a new menu item, though adding this link html to menu.html is never discussed.

Hence, add the line at 4 to existing menu.html to dynamically create a "Create New Item" link for each visited restaurant page, which matches with the video at L3-11.

Additionally, the menu.html from L3-10 does not generate html whose formatting matches that shown in the video.  Accordingly, replace <p></p> usage with a few </br> as discussed in the video. As such, the formatting of this menu.html will not match the menu.html in the L3-10 folder, but it should more closely match the formatting shown on the video.