Use consistent routing variables
As distributed, project.py for this quiz renders the editmenuitem.html template with the chosen parameter names "restaurant_id", "MenuID", and "item". As such, editmenuitem.html is constrained to using those variable names to access the respective objects, and throws an error when trying to obtain "menu_id" in line 3 (indeed, there is no "menu_id").

An intuitive solution would be to edit line 3 of `editmenuitem.html` so that it looks for "MenuID" (as discussed in https://github.com/lobrown/Full-Stack-Foundations/pull/62).

However, menu.html also expects the variable to be named "menu_id".

Stylistically, rather than changing all instances of the previously-used "menu_id" to the apparently novel "MenuID", simply change project.py to consistently use "menu_id" internally and when passing parameters to templates (e.g. `editmenuitem.html`).